refactor(envelope): split capability owner
